+++ mitxgraders/baseclasses.py
@@ -13,8 +13,8 @@ Also contains some error classes:
 from __future__ import division
 import numbers
 import abc
-from mitxgraders.voluptuous import Schema, Required, All, Any, Range, MultipleInvalid
-from mitxgraders.voluptuous.humanize import validate_with_humanized_errors as voluptuous_validate
+from voluptuous import Schema, Required, All, Any, Range, MultipleInvalid
+from voluptuous.humanize import validate_with_humanized_errors as voluptuous_validate
 from mitxgraders.version import __version__
 
 class ConfigError(Exception):
