+++ mitxgraders/formulagrader.py
@@ -11,11 +11,11 @@ from collections import namedtuple
 from pprint import PrettyPrinter
 import re
 import itertools
+from voluptuous import Schema, Required, Any, All, Extra, Invalid, Length
 from mitxgraders.sampling import (VariableSamplingSet, FunctionSamplingSet, RealInterval,
                                   DiscreteSet, gen_symbols_samples, construct_functions,
                                   construct_constants, construct_suffixes)
 from mitxgraders.baseclasses import ItemGrader, InvalidInput, ConfigError
-from mitxgraders.voluptuous import Schema, Required, Any, All, Extra, Invalid, Length
 from mitxgraders.helpers.calc import CalcError, evaluator, parsercache
 from mitxgraders.helpers.validatorfuncs import (Positive, NonNegative, is_callable,
                                                 PercentageString, is_callable_with_args, all_unique)
