+++ pytest.ini
@@ -1,4 +1,3 @@
 [pytest]
-norecursedirs = mitxgraders/voluptuous
 addopts = --doctest-modules --cov=mitxgraders --cov-report=term-missing
 testpaths = tests mitxgraders
